[REF] mail: introduce 'mail.out_of_focus' service
Isolate out of focus concern of messaging in its own service
